Enable editing for course details
