+++ app/templates/components/course-editing.hbs
@@ -1,4 +1,4 @@
-{{course-overview course=course editableDirectors=true}}
+{{course-overview course=course editing=true}}
 {{detail-objectives objectives=course.objectives}}
 {{detail-learning-materials materials=course.learningMaterials}}
 {{detail-competencies competencies=course.competencies}}
